Added Members Repo; added new getWhere function to Repo.java (fits SQLService better); Updated AdminController to use new MembersRepo

+++ XYZDrivers/src/main/java/com/xyzdrivers/models/Member.java
@@ -49,7 +49,7 @@ public class Member extends Model
     public String getStatus() {
         return status;
     }
-    public float getBalance() {
+    public double getBalance() {
         return balance;
     }
 }
